+++ src/assets/scss/_utilities.scss
@@ -144,7 +144,7 @@
   
 
 
-  /*REACT WIDGETS: multiselect*/
+/*REACT WIDGETS: multiselect*/
 button.rw-multiselect-tag-btn.rw-btn.rw-btn-select {
   background: transparent;
   border: 0px;
