+++ src/containers/index.js
@@ -13,3 +13,5 @@ export { default as MyHeadersPageContainer } from './MyHeadersPageContainer';
 export { default as EditHeaderPageContainer } from './EditHeaderPageContainer';
 export { default as DocumentHeaderContainer } from './DocumentHeaderContainer';
 export { default as PricingPageContainer } from './PricingPageContainer';
+export { default as CreateDocumentWithQuestionFormContainer } from './CreateDocumentWithQuestionFormContainer';
+
